display validation on each step
